requirements.txt
# requirements.txt

vk_api
requests~=2.32.3
beautifulsoup4
rapidfuzz
pathlib
python-dotenv~=1.1.0
APScheduler~=3.11.0